refactor(queue): extract signal-snapshot generation into its own module
Part of #4013's module-split sequence (step 2, after transient-locks.ts
in #4157): generateSignalSnapshots and its loadOpenQueueCounts helper
move to src/queue/signal-snapshot.ts. Pure mechanical move, no behavior
change -- a re-export shim keeps processors.ts's existing external
callers (src/api/routes.ts, test/unit/queue-trends.test.ts) working
unchanged.
